refactor(desktop): extract ClaudeSubscriptionCard into its own file
Refs #1042. The Claude paste-code subscription card was a ~400-line
sub-component inside provider-oauth-section.tsx with its own five-action
pending guard and experimental gate. Pure move into
claude-subscription-card.tsx (card + ClaudeSubscriptionPendingAction +
SubscriptionStatePresentation + presentSubscriptionState); the OAuth
section imports it back.

The provider contract-source helper gains the new file (kept right after
the OAuth section in the combined source so cross-file span pins keep
matching), and the chip-converge / web-search-boundary file lists follow
the move.
